Lot of sprite work
Adding animations
- knocked down
- ko
- get up

Also adding another skin tone and hair colors. And to think I'll have to do this for women too. Just, a lot of work. Probably 10 hours total just on this worker sprite, BUT, the worker is effectively the most common character in the game, so this is like most of the scenery.